fix header types

+++ include/rule_2d.h
@@ -12,7 +12,7 @@ class c_rule_2d
     public:
         c_rule_2d(uint16_t rule);
         void set_rule(uint16_t rule);
-        uint8_t get_rule_case(float left_cell, float center_cell, float right_cell) const;
+        uint8_t get_rule_case(uint8_t left_cell, uint8_t center_cell, uint8_t right_cell) const;
         void print_rule() const;
     private:
         e_rule_type m_rule_type;
